Changed the generator to match the database
Updated the generator to ensure that it matches the data base that is
being created. Still needs work on the voca_clarification but that is
due to me not really knowing what data should go into that slot. The
date of birth is still currently set to adults only, but that will
change once we figure out more of what is needed to fit the OUHSC
departments needs.