Got the database creation working
The database is able to be created using the code provided. I will later
write out documentation for all of this, but for now I want to see if I
can get it all working on all developers computers. The database will
still probably need to be put online, but we also now have the option to
just have it run locally. Various of features are also pushed in this
commit, but all have to do with the database creating tables correctly.