Implementation of 3 more tables
Created 3 more tables to be generated by the data-generator.py file.
Please note that the cac_case is still not done yet as I need to think
of a way to generate some of the data. I also added a util.py file to
clean up some of the main source code and make the overall readability
much better.