Got button linking working for 2 interfaces
shows an example of how to get from one interface to the next interface
so we can get all the interfaces linked. Notice the __init__ calls in
all functions, as well as changing the root to self in each interface.
Also ensure the scrollbar and canvas are not pack but grid