Skeleton iteration of populating_database.py done
The functionality for populating the databases is now done. We just need
to go through and start populating each one with the correct
information. Should just be .sql file work from here, but im sure we
will run into more errors. Also the use of panda will be required again.